back/pyproject.toml
[tool.poetry]
name = "grammar-demo"
version = "0.1.0"
description = ""
authors = ["Evgeny Nizhibitsky <nizhib@github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
revchatgpt = "^2.3.0"
uvicorn = "^0.20.0"
fastapi = "^0.92.0"
pydantic = "^1.10.5"
python-dotenv = "^0.21.1"
ujson = "^5.7.0"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
